Added Front-end Framework Code
I added the JavaScript code which contains a the front-end section of the application, the framework used is BackboneJS along with RequireJS for dependency indection. This consumes the data from the backend. Consequently, this section contains a numbers of views, models and controllers. Lastly, MustacheJS has been used to provide templates for the views.